disallow deletion of notes for trainees

+++ dak/process_new.py
@@ -703,7 +703,7 @@ def do_new():
             edit_note(database.get_new_comments(changes.get("source", "")))
         elif answer == 'P' and not Options["Trainee"]:
             prod_maintainer(database.get_new_comments(changes.get("source", "")))
-        elif answer == 'R':
+        elif answer == 'R' and not Options["Trainee"]:
             confirm = utils.our_raw_input("Really clear note (y/N)? ").lower()
             if confirm == "y":
                 database.delete_new_comments(changes.get("source"), changes.get("version"))
